backend/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15
django-cors-headers>=4.0
gunicorn>=21.0
whitenoise>=6.5
psycopg2-binary>=2.9
